chore(refactor): move components to common
