+++ src/common/components/JsonForm/CustomWidgets/LocationSelector/LocationSelector.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+interface Props {
+  value: string
+  onChange: (value: string) => void
+}
+
+const LocationSelector: React.FunctionComponent<Props> = ({
+  value,
+  onChange,
+}) => {
+  return <div />
+}
+
+export default LocationSelector
