+++ src/common/components/SingleStatistic/SingleStatistic.styles.ts
