+++ src/common/components/SingleStatistic/SingleStatistic.tsx
