+++ src/common/routing/Routes.tsx
@@ -1,21 +1,21 @@
 import * as React from 'react'
 import { Route, Switch } from 'react-router-dom'
-import { ProjectContainerConnected } from './project/ProjectContainer'
-import { EntitiesContainerConnected } from '../modules/Entities/Entities.container'
-import { EntitiesSelectConnected } from '../modules/Entities/Entities.container.select'
-import { ProjectCreateConnected } from './project/ProjectCreate'
-import { contentType } from '../types/models'
-import { ProjectForm } from './project/ProjectForm'
+import { ProjectContainerConnected } from '../../components/project/ProjectContainer'
+import { EntitiesContainerConnected } from '../../modules/Entities/Entities.container'
+import { EntitiesSelectConnected } from '../../modules/Entities/Entities.container.select'
+import { ProjectCreateConnected } from '../../components/project/ProjectCreate'
+import { contentType } from '../../types/models'
+import { ProjectForm } from '../../components/project/ProjectForm'
 import { Fragment } from 'react'
-import { RegisterConnected } from './register/RegisterContainer'
-import { NotFound } from './public/NotFound'
-import { Spinner } from './common/Spinner'
-import { ProjectSignAndCreateConnected } from './project/curation/ProjectSignAndCreate'
-import { ProjectCreateUploadPublicDocsConnected } from './project/curation/ProjectCreateUploadPublicDocs'
-import { Overview } from '../common/components/Bonds/OverviewWrapper/Overview'
-import Exchange from '../common/components/Bonds/ExchangeWrapper/Exchange'
-import Orders from '../modules/BondAccountOrders/BondAccountOrders.container'
-import { UnderConstruction } from './public/UnderConstruction'
+import { RegisterConnected } from '../../components/register/RegisterContainer'
+import { NotFound } from '../../components/public/NotFound'
+import { Spinner } from '../components/Spinner'
+import { ProjectSignAndCreateConnected } from '../../components/project/curation/ProjectSignAndCreate'
+import { ProjectCreateUploadPublicDocsConnected } from '../../components/project/curation/ProjectCreateUploadPublicDocs'
+import { Overview } from '../components/Bonds/OverviewWrapper/Overview'
+import Exchange from '../components/Bonds/ExchangeWrapper/Exchange'
+import Orders from '../../modules/BondAccountOrders/BondAccountOrders.container'
+import { UnderConstruction } from '../../components/public/UnderConstruction'
 import { SubmitEntityClaimConnected } from 'src/modules/SubmitEntityClaim/SubmitEntityClaim.container'
 
 export const Routes: React.SFC<{}> = props => {
