+++ src/components/project/ProjectAgents.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { WidgetWrapper } from '../common/WidgetWrapper'
-import { LayoutWrapper } from '../common/LayoutWrapper'
-import { ButtonTypes, Button } from '../common/Buttons'
+import { WidgetWrapper } from '../../common/components/Wrappers/WidgetWrapper'
+import { LayoutWrapper } from '../../common/components/Wrappers/LayoutWrapper'
+import { ButtonTypes, Button } from '../../common/components/Form/Buttons'
 
 const Heading = styled.h2`
   color: white;
