+++ src/components/project/ProjectContainer.tsx
@@ -19,7 +19,7 @@ import { ProjectSingleClaim } from './ProjectSingleClaim'
 import { ProjectClaims } from './ProjectClaims'
 import styled from 'styled-components'
 import { ProjectAgents } from './ProjectAgents'
-import { Spinner } from '../common/Spinner'
+import { Spinner } from '../../common/components/Spinner'
 import { UserInfo } from '../../modules/Account/types'
 import { ProjectSidebar } from './ProjectSidebar'
 import * as Toast from '../../common/utils/Toast'
