+++ src/components/project/ProjectNewClaim.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
-import { LayoutWrapperClaims } from '../common/LayoutWrapperClaims'
-import { WidgetWrapperClaims } from '../common/WidgetWrapperClaims'
+import { LayoutWrapperClaims } from '../../common/components/Wrappers/LayoutWrapperClaims'
+import { WidgetWrapperClaims } from '../../common/components/Wrappers/WidgetWrapperClaims'
 import DynamicForm from '../../common/components/Form/DynamicForm/DynamicForm'
 import { decode as base64Decode } from 'base-64'
 import { Data } from '../../modules/project/types'
 import styled from 'styled-components'
 import { FormStyles } from '../../types/models'
-import { Spinner } from '../common/Spinner'
+import { Spinner } from '../../common/components/Spinner'
 import { successToast } from '../../common/utils/Toast'
 
 const FormContainer = styled.div`
