+++ src/components/project/funding/FundingContainer.tsx
@@ -8,9 +8,9 @@ import { Web3Acc } from 'src/modules/web3/web3'
 import { FundingGauge } from './FundingGauge'
 import { FundingButton } from './FundingButton'
 import { Fragment } from 'react'
-import { ModalWrapper } from 'src/components/common/ModalWrapper'
+import { ModalWrapper } from 'src/common/components/Wrappers/ModalWrapper'
 import { successToast, errorToast } from '../../../common/utils/Toast'
-import { Button, ButtonTypes } from 'src/components/common/Buttons'
+import { Button, ButtonTypes } from 'src/common/components/Form/Buttons'
 import { BigNumber } from 'bignumber.js'
 import { Header } from '../../../types/models'
