+++ src/components/project/overview/modalContent/NoKYC.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { ButtonTypes, Button } from 'src/components/common/Buttons'
+import { ButtonTypes, Button } from 'src/common/components/Form/Buttons'
 import { getIxoWorldRoute } from 'src/common/utils/formatters'
 
 const BorderBox = styled.div`
