+++ src/components/project/overview/modalContent/NoKeysafe.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { ButtonTypes, Button } from 'src/components/common/Buttons'
+import { ButtonTypes, Button } from 'src/common/components/Form/Buttons'
 
 const BorderBox = styled.div`
   border: 1px solid #004c61;
