+++ src/components/project/overview/modalContent/NotLedgered.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { ButtonTypes, Button } from 'src/components/common/Buttons'
+import { ButtonTypes, Button } from 'src/common/components/Form/Buttons'
 import Success from '../../../../assets/icons/Success'
 
 const ModalData = styled.div`
