+++ src/components/register/RegisterContainer.tsx
@@ -4,14 +4,14 @@ import { UserInfo } from '../../modules/Account/types'
 import { connect } from 'react-redux'
 import { RootState } from '../../common/redux/types'
 import { errorToast, successToast } from '../../common/utils/Toast'
-import { ModalWrapper } from '../common/ModalWrapper'
+import { ModalWrapper } from '../../common/components/Wrappers/ModalWrapper'
 import { AgentRoles } from '../../types/models'
 import { Banner } from './Banner'
 import { TextBlock } from './TextBlock'
 import { deviceWidth } from '../../lib/commonData'
 import MediaQuery from 'react-responsive'
-import { Button, ButtonTypes } from '../common/Buttons'
-import { Spinner } from '../common/Spinner'
+import { Button, ButtonTypes } from '../../common/components/Form/Buttons'
+import { Spinner } from '../../common/components/Spinner'
 import { Header } from '../../types/models'
 import Kyc from '../../assets/icons/Kyc'
 import Claims from '../../assets/icons/Claims'
