+++ src/modules/App/App.tsx
@@ -11,10 +11,10 @@ import { initIxo } from '../ixo/ixo.actions'
 import { initKeysafe } from '../keysafe/keysafe.actions'
 import { UserInfo } from '../Account/types'
 import { updateLoginStatus } from '../Account/Account.actions'
-import ScrollToTop from '../../components/common/ScrollToTop'
+import ScrollToTop from '../../common/components/ScrollToTop'
 import '../../assets/icons.css'
-import { Routes } from '../../components/Routes'
-import { Spinner } from '../../components/common/Spinner'
+import { Routes } from '../../common/routing/Routes'
+import { Spinner } from '../../common/components/Spinner'
 import { ToastContainer } from 'react-toastify'
 import * as ReactGA from 'react-ga'
 import { connectWeb3 } from '../web3/web3.actions'
